GitHub_PUBLIC/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
bottleneck=1.3.7=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py310hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=py310haa95532_0
cffi=1.16.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.0.5=py310h59b6b97_0
cpuonly=2.0=0
cycler=0.11.0=pyhd3eb1b0_0
cython=3.0.10=pypi_0
debugpy=1.8.1=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
et_xmlfile=1.1.0=py310haa95532_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=py310haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
future=1.0.0=pypi_0
gmpy2=2.1.2=py310h7f96b67_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.7=py310haa95532_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.24.0=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=py310haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h2bbff1b_1
json5=0.9.25=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.4=py310hd77b12b_0
krb5=1.20.1=h5b6d351_1
latexcodec=3.0.0=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
libxgboost=1.7.3=hd77b12b_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib=3.8.0=py310haa95532_0
matplotlib-base=3.8.0=py310h4ed8f06_0
matplotlib-inline=0.1.7=pypi_0
mendeleev=0.16.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
monty=2024.5.24=pypi_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
multidict=6.0.5=pypi_0
munkres=1.1.4=py_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=py310haa95532_0
notebook=7.2.0=pypi_0
notebook-shim=0.2.4=pypi_0
numexpr=2.8.4=py310hd213c9f_0
numpy=1.26.4=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.2=py310h2bbff1b_0
openssl=3.0.13=h2bbff1b_2
overrides=7.7.0=pypi_0
packaging=23.2=py310haa95532_0
palettable=3.3.3=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pillow=10.3.0=py310h2bbff1b_0
pip=24.0=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
plotly=5.22.0=pypi_0
ply=3.11=py310haa95532_0
pooch=1.7.0=py310haa95532_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.45=pypi_0
psutil=5.9.8=pypi_0
pure-eval=0.2.2=pypi_0
py-xgboost=1.7.3=py310haa95532_0
pybtex=0.24.0=pypi_0
pycparser=2.22=pypi_0
pyfiglet=0.8.post1=pypi_0
pyg-lib=0.4.0+pt23cpu=pypi_0
pygments=2.18.0=pypi_0
pymatgen=2024.5.1=pypi_0
pyparsing=3.1.2=pypi_0
pyqt=5.15.10=py310hd77b12b_0
pyqt5-sip=12.13.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.14=he1021f5_1
python-dateutil=2.9.0post0=py310haa95532_2
python-json-logger=2.0.7=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.3.0=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=py310haa95532_0
pywin32=306=pypi_0
pywinpty=2.0.13=pypi_0
pyyaml=6.0.1=py310h2bbff1b_0
pyzmq=26.0.3=pypi_0
qt-main=5.15.2=h19c9488_10
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
rdkit=2023.9.6=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.2=py310haa95532_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.18.1=pypi_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=69.5.1=py310haa95532_0
sip=6.7.12=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
spglib=2.4.0=pypi_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py310haa95532_0
tabulate=0.9.0=pypi_0
tbb=2021.12.0=pypi_0
tenacity=8.3.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h0416ee5_0
tomli=2.0.1=py310haa95532_0
torch-cluster=1.6.3+pt23cpu=pypi_0
torch-geometric=2.5.3=pypi_0
torch-geometric-temporal=0.54.0=pypi_0
torch-scatter=2.1.2+pt23cpu=pypi_0
torch-sparse=0.6.18+pt23cpu=pypi_0
torch-spline-conv=1.2.2+pt23cpu=pypi_0
torchaudio=2.3.0=py310_cpu
torchvision=0.18.0=py310_cpu
tornado=6.3.3=py310h2bbff1b_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.11.0=py310haa95532_0
tzdata=2024a=h04d1e81_0
uncertainties=3.1.7=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=py310haa95532_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=py310haa95532_0
widgetsnbextension=4.0.11=pypi_0
win_inet_pton=1.1.0=py310haa95532_0
xlrd=2.0.1=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
